pyproject.toml
[project]
name = "autogen-yt-agent"
version = "0.1.0.dev1"
description = "A YouTube agent for AutoGen"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyright==1.1.389",
    "mypy==1.13.0",
    "ruff==0.4.8",
    "autogen-agentchat==0.4.0.dev8",
    "openai-whisper",
    "ffmpeg-python",
    "yt-dlp",
    "opencv-python",
    "pip>=24.3.1",
]

[tool.ruff]
include = ["src/**"]
line-length = 120
fix = true

[tool.ruff.format]
docstring-code-format = true

[tool.uv]
prerelease = "allow"
